Added Image Added pixel format support Added MemoryStream Added Rect Added ResourceLoader Added generic loader (bmp, gif, hdr, jpg, jpeg, pic, png, psd, tga) Added PCX loader Added utility module initializer Fixed Config.hpp include Prerequesites.hpp now overwrites _WIN32_WINNT when defined version is less than requiered version Renderer's initialisation will implicitly initialize utility module Removed RENDERER_SINGLETON option Shaders are now resources

+++ include/Nazara/Math/Matrix4.hpp
@@ -106,8 +106,7 @@ template<typename T> class NzMatrix4
 			T m31, m32, m33, m34;
 			T m41, m42, m43, m44;
 
-			unsigned int refCount;
-
+			unsigned short refCount;
 			NazaraMutex(mutex)
 		};
